Game.js can be included on any page with multiple game canvases
This commit allows multiple instances of Game to be created on a
single page, with each being shown on a different canvas.  Also, the
js dependencies of game.js are now loaded dynamically.  This allows
anybody to include game.js on any page with a canvas with class
'layout' and show a game canvas.

Code still needs to be added to add a saved layout file as data source
and show that game in the canvas.